Upload creates artists and albums
